requirements.txt
# Core dependencies
streamlit>=1.37.0
streamlit-option-menu>=0.3.6
pandas>=2.0.0
plotly>=5.17.0
pydantic==2.5.0
python-dotenv==1.0.0

# Data ingestion
feedparser==6.0.10
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
watchdog==3.0.0

# Embeddings and ML
google-generativeai>=0.8.3
google-genai>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.8
numpy==1.24.3
scikit-learn==1.3.2

# Image Generation
pillow>=10.0.0
requests>=2.31.0

# Vector stores
faiss-cpu==1.7.4
pathway-client==0.1.0

# Text processing
beautifulsoup4==4.12.2
pypdf==3.17.0
python-magic==0.4.27
chardet==5.2.0

# Utilities
requests==2.31.0
aiohttp==3.9.0
asyncio-throttle==1.0.2
schedule==1.2.0

# Logging and monitoring
structlog==23.2.0
rich==13.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Optional: Jupyter for experimentation
jupyter==1.0.0
ipykernel==6.26.0